+++ services/disk/disk_contract.hpp
@@ -94,7 +94,7 @@ namespace services::disk {
         storage_scan(session_id_t session,
                      collection_full_name_t name,
                      std::unique_ptr<components::table::table_filter_t> filter,
-                     int limit,
+                     int64_t limit,
                      components::table::transaction_data txn);
         actor_zeta::unique_future<std::unique_ptr<components::vector::data_chunk_t>>
         storage_scan_projected(session_id_t session,
