+++ services/disk/manager_disk.cpp
@@ -1039,7 +1039,7 @@ namespace services::disk {
     manager_disk_t::storage_scan(session_id_t /*session*/,
                                  collection_full_name_t name,
                                  std::unique_ptr<components::table::table_filter_t> filter,
-                                 int limit,
+                                 int64_t limit,
                                  components::table::transaction_data txn) {
         auto* s = get_storage(name);
         if (!s) {
@@ -1628,7 +1628,7 @@ namespace services::disk {
     manager_disk_empty_t::storage_scan(session_id_t /*session*/,
                                        collection_full_name_t name,
                                        std::unique_ptr<components::table::table_filter_t> filter,
-                                       int limit,
+                                       int64_t limit,
                                        components::table::transaction_data txn) {
         auto* s = get_storage(name);
         if (!s) {
