+++ services/disk/manager_disk.hpp
@@ -194,7 +194,7 @@ namespace services::disk {
         storage_scan(session_id_t session,
                      collection_full_name_t name,
                      std::unique_ptr<components::table::table_filter_t> filter,
-                     int limit,
+                     int64_t limit,
                      components::table::transaction_data txn);
         // Projected scan: returns only the first column_limit columns. Pre-allocates result.
         unique_future<std::unique_ptr<components::vector::data_chunk_t>>
@@ -485,7 +485,7 @@ namespace services::disk {
         storage_scan(session_id_t session,
                      collection_full_name_t name,
                      std::unique_ptr<components::table::table_filter_t> filter,
-                     int limit,
+                     int64_t limit,
                      components::table::transaction_data txn);
         unique_future<std::unique_ptr<components::vector::data_chunk_t>>
         storage_scan_projected(session_id_t session,
